requirements.txt
plotly==4.14.3
streamlit==0.80.0
joblib==0.17.0
matplotlib==3.3.2
pandas==1.1.3
seaborn==0.11.0
numpy==1.19.2
mysql_connector_repackaged==0.3.1
scikit_learn==0.24.1
